Added module locally, changing project to full front end web application
I manage to add a module to my javascript project locally. By
literally downloading the module and adding to it to my project.
I believe that because this is actually pretty far to how the industry
actually creates web applications that I am going to change my
project to resemble no matter how little to an actual web application.

This is me just currently saving my progress, I expect that everything
already is pretty broken.